app/cart/requirements.txt
alembic==0.8.2
blinker==1.4
Flask==0.10.1
Flask-Admin==1.3.0
Flask-Login==0.3.0
Flask-Mail==0.9.1
Flask-Migrate==1.5.1
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==1.7.4
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.2
MarkupSafe==0.23
passlib==1.6.5
prices==0.5.4
python-editor==0.4
python-slugify==1.1.3
satchless==1.1.3
smartencoding==0.3
SQLAlchemy==1.0.8
Unidecode==0.4.18
Werkzeug==0.10.4
WTForms==2.0.2
